+++ R/gseaplot.R
@@ -184,7 +184,7 @@ hplot <- function(x, geneSetID) {
         stop("hplot only work for GSEA result")
     }
 
-    check_installed('ggHoriPlot', 'for `hplot()`.')
+    require_suggested('ggHoriPlot', 'for `hplot()`.')
 
     gsdata <- get_gsdata(x, geneSetID)